ini_parse4: remove strArray from class member
Previously strArray is used to generate the ini file.

Now, we maintain the hash table "htProfile" which contains the section, key and
value, use this hash table to generate the init file.
Therefore strArray is no more needed as a class member, and just used as a
local variables for Open() function.